requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
Authlib==1.6.0 
redis
python-dotenv
redis==4.6.0
requests-oauthlib==1.3.1
requests==2.31.0
Flask-Migrate==3.1.0
psycopg2-binary==2.9.5
